Field hiding fix + element grab cleanup
Elements weren't revealed if some data is already present in the form or
if the country is changed away from UK.

The ID for the search-container was grabbed via standard '#' symbol with
jQuery. This caused an issue since sometimes the passed id (from name)
contained brackets. Now using attribute selector with quotemarks.
Disabled not implemented functionality configurations.